requirements.txt
django-extensions==3.2.3
unidecode==1.3.8
markdown==3.7